+++ src/lib/models/guild.js
@@ -42,21 +42,19 @@ module.exports = (sequelize, DataTypes) => {
         key: 'id',
       },
     },
-  },
-  // {
-  //   classMethods: {
-  //     associate (models) {
-  //       Gw2Guild.belongsTo(models.Gw2ApiToken, {
-  //         onDelete: 'SET NULL',
-  //         foreignKey: {
-  //           name: 'apiTokenId',
-  //           allowNull: false,
-  //         },
-  //       });
-  //     },
-  //   },
-  // }
-  );
+  }, {
+    classMethods: {
+      associate (models) {
+        Gw2Guild.belongsTo(models.Gw2ApiToken, {
+          onDelete: 'SET NULL',
+          foreignKey: {
+            name: 'apiTokenId',
+            allowNull: false,
+          },
+        });
+      },
+    },
+  });
 
   return Gw2Guild;
 };