+++ CHANGELOG.md
@@ -1,12 +1,30 @@
-# Change Log
+# Changelog
 
-## Log structure
+All notable changes to this project will be documented in this file.
 
-```
-## [Unreleased|major.minor.patch] - yyyy/mm/dd
-### Added|Changed|Fixed|Removed
-- Changes
-```
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
+and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [Unreleased][]
+### Changed
+- Babel configuration to use `babel-preset-env`
+- Docker images to use node `7.8.0`
+- Leaderboard pagination to match everything else
+- Guild members to be orded by armory users first
+- Gw2 api fetch to cache in memory for 5 minutes
+
+### Removed
+- Characters from base guild resource
+- Characters from user resource
+- Users from base guild resource
+
+### Added
+- guild/characters resource
+- guild/users resource
+- Characters/guilds of the day
+
+### Fixed
+- Guild resource to return 404 if not found
 
 ## [2.4.0] - 2017/04/05
 ### Added